js/index-1.ts: seamless autorefresh support
This patch adds seamless autorefresh. Every 60 seconds, the client will re-fetch the statistics data. Because the table is built in the background, it won't be modified until it's completely ready - all the browser has to do, essentially, is re-parse the HTML once it's pushed from generate_table into the DOM, a relatively fast operation compared to the other JS operations which have to happen for the table to be recreated.

Signed-off-by: Amy Parker <amy@amyip.net>