allow custom setting of thread count
This patch adds support for custom setting of thread count. Support is not being added to NIx/via the config file, because Nix and other platforms should be using sane defaults. If there's ever a desire to add support to Nix, because of envvar usage, doing so would be trivial.

This patch also sets the default number of threads for interval.py to 4*nproc instead of the static 8.

Fixes: #38

Signed-off-by: Amy Parker <amy@amyip.net>